Move component to container dir
