+++ frontend/src/metabase/dashboard/containers/DashboardHeader.styled.jsx
